fix: pass only provider class name, not an object
